requirements.txt
contextlib2==0.5.5
decorator==4.3.0
enum34==1.1.6
graphviz==0.9
h5py==2.8.0
joblib==0.12.4
networkx==2.1
numpy==1.15.1
pandas==0.23.4
patsy==0.5.0
pymc3==3.5
pyro-ppl==0.2.1
python-dateutil==2.7.3
pytz==2018.5
scipy==1.1.0
six==1.11.0
Theano==1.0.2
torch==0.4.1
tqdm==4.25.0
